Fix issues with complete signup mutation
